Cleaned up original gsf_read file uploaded by Carlo
